+++ src/Client/TrapHandle.php
@@ -16,14 +16,10 @@ use Symfony\Component\VarDumper\Caster\TraceStub;
 final class TrapHandle
 {
     private bool $haveToSend = true;
-
     private int $times = 0;
-
     private string $timesCounterKey = '';
-
     private int $depth = 0;
-
-    private StaticState $staticState;
+    private readonly StaticState $staticState;
 
     private function __construct(
         private array $values,
@@ -139,7 +135,7 @@ final class TrapHandle
      */
     public function return(int|string $key = 0): mixed
     {
-        if (\count($this->values) === 0) {
+        if ($this->values === []) {
             throw new \InvalidArgumentException('No values to return.');
         }
 
@@ -176,8 +172,6 @@ final class TrapHandle
      * ```php
      * trap()->context(['foo bar', => 42, 'baz' => 69]);
      * ```
-     *
-     * @param mixed ...$values
      */
     public function context(mixed ...$values): self
     {