+++ src/Command/Run.php
@@ -31,9 +31,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 final class Run extends Command implements SignalableCommandInterface
 {
     private ?Application $app = null;
-
     private Logger $logger;
-
     private bool $cancelled = false;
 
     public function configure(): void
@@ -60,6 +58,7 @@ final class Run extends Command implements SignalableCommandInterface
      */
     public function getServers(Container $container): array
     {
+        /** @var TcpPorts $config */
         $config = $container->get(TcpPorts::class);
 
         $servers = [];
@@ -73,7 +72,7 @@ final class Run extends Command implements SignalableCommandInterface
             $port > 0 && $port < 65536 or throw new \InvalidArgumentException(
                 \sprintf('Invalid port `%s`. It must be in range 1-65535.', $port),
             );
-            $servers[] = new SocketServer($port, $config->host, $config->type);
+            $servers[] = new SocketServer($port, $config->host, $config->type, $config->pollingInterval);
         }
         return $servers;
     }