+++ src/Sender/RemoteSender.php
@@ -13,11 +13,10 @@ use Buggregator\Trap\Support\Uuid;
 final class RemoteSender extends SocketSender
 {
     private string $uuid;
-
     private readonly string $clientVersion;
 
     public function __construct(
-        string $uuid = null,
+        ?string $uuid = null,
         string $host = '127.0.0.1',
         int $port = 9912,
     ) {