+++ src/Socket/Client.php
@@ -18,14 +18,11 @@ final class Client implements Destroyable
     /** @var string[] */
     private array $writeQueue = [];
 
-    /** @var string */
     private string $readBuffer = '';
-
     private bool $toDisconnect = false;
-
     private \Closure $onPayload;
-
     private \Closure $onClose;
+    private Timer $selectTimer;
 
     /**
      * @param positive-int $payloadSize
@@ -33,7 +30,9 @@ final class Client implements Destroyable
     private function __construct(
         private readonly \Socket $socket,
         private readonly int $payloadSize,
+        float $selectPeriod,
     ) {
+        $this->selectTimer = new Timer($selectPeriod);
         \socket_set_nonblock($this->socket);
         $this->setOnPayload(static fn(string $payload) => null);
         $this->setOnClose(static fn() => null);
@@ -41,12 +40,18 @@ final class Client implements Destroyable
 
     /**
      * @param positive-int $payloadSize Max payload size.
+     * @param float $selectPeriod Time to wait between socket_select() calls in seconds.
      */
     public static function init(
         \Socket $socket,
         int $payloadSize = 10485760,
+        float $selectPeriod = .001,
     ): self {
-        return new self($socket, $payloadSize);
+        return new self(
+            socket: $socket,
+            payloadSize: $payloadSize,
+            selectPeriod: $selectPeriod,
+        );
     }
 
     public function destroy(): void
@@ -103,6 +108,7 @@ final class Client implements Destroyable
                 throw new ClientDisconnected();
             }
             \Fiber::suspend();
+            $this->selectTimer->reset()->wait();
         } while (true);
     }