+++ CHANGELOG.md
@@ -1,5 +1,41 @@
 # Changelog
 
+## 1.11.1 (2024-12-07)
+
+## What's Changed
+* Resolve PHP 8.4 deprecations by @roxblnfk in https://github.com/buggregator/trap/pull/144
+
+
+**Full Changelog**: https://github.com/buggregator/trap/compare/1.11.0...1.11.1
+
+## 1.11.0 (2024-09-25)
+
+## What's Changed
+* Add config for main loop interval and socket polling interval by @roxblnfk in https://github.com/buggregator/trap/pull/141
+
+
+**Full Changelog**: https://github.com/buggregator/trap/compare/1.10.2...1.11.0
+
+## 1.10.2 (2024-09-24)
+
+## What's Changed
+* Fix Psalm issues by @Kaspiman in https://github.com/buggregator/trap/pull/131
+* Add Rector to CI by @Kaspiman in https://github.com/buggregator/trap/pull/133
+* Separate polling intervals for sockets by @roxblnfk in https://github.com/buggregator/trap/pull/140
+
+## New Contributors
+* @Kaspiman made their first contribution in https://github.com/buggregator/trap/pull/131
+
+**Full Changelog**: https://github.com/buggregator/trap/compare/1.10.1...1.10.2
+
+## 1.10.1 (2024-06-23)
+
+## What's Changed
+* Better HTTP/SMTP Multipart parsing by @roxblnfk in https://github.com/buggregator/trap/pull/128
+
+
+**Full Changelog**: https://github.com/buggregator/trap/compare/1.10.0...1.10.1
+
 ## 1.10.0 (2024-06-20)
 
 ## What's Changed